Doc: Document limitations of QScreenCapture
Details: Added known limitations in QScreenCapture class
and the ScreenCapture QML type references.

As the section is identical for both (except for the type name),
use the \include command to pull in the content from a .qdocinc
file. For C++ reference, prefix the type name with 'Q'.

Add [Q]WindowCapture as a see-also entry.

Pick-to: 6.6
Task-number: QTBUG-117407
Change-Id: I1e94fd7d6af04026306fe6fc86aa095ba4e6f376
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>