Fix next/previous buttons issue in Media Player app example
If we loaded several videos, clicking on the next button
moved on to the next video, but if we did this on the last video,
the app would select an 'empty' video and only on the second click
would it move to the front of the list.
The same was happening with the previous button.

The problem was the function that determines
the index of the next preceding video, so I fixed it.

Fixes: QTBUG-119083
Pick-to: 6.6
Change-Id: I6fc573cfc09338c8c8dc4aef7f8d0b8a6e74290e
Reviewed-by: Lars Knoll <lars@knoll.priv.no>