Improve the criteria of the demuxer's bufferization
AVPacket::duration can be empty and should also rely on the position
the difference when checking stream buffering limitations.

In the previous implementation, we calculated "bufferedDuration" for
each stream as an amount of AVPacket::duration for all packets that
have been got from the ffmpeg demuxer but haven't been processed
(decoded).

In some cases, ffmpeg exposes packets with duration = 0 that break
the mechanism. The CR introduces a new metric
"maxSentPacketPos - maxProcessedPacketPos" that shows approximately
the same. Probably, the previous criteria will be dropped after
additional investigations.

Pick-to: 6.6 6.5
Change-Id: Ic138ea267080a65a969bc77be52e50d3bb96030a
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>