Add 'autoPlay' property to qml API
The property 'autoPlay' used to be available in Qt5,
and was removed in Qt6. Some users want this property as
it makes them able to follow qml declarative style
instead of calling play() in handlers.

The property makes sense only in qml, let's not add it to c++ API.

Task-number: QTBUG-110312
Change-Id: Ic060fab3dba11155d34343bcec04685fa15d19c4
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>