EVRCustomPresenter: remove redundant "repaint" logic
Remove the code that handles the "repaint" logic and that was originally
taken from the standard EVR Presenter example
https://learn.microsoft.com/en-us/windows/win32/medfound/how-to-write-an-evr-presenter

It's not used currently and perhaps was never used before.
Also it doesn't seem as something that could be used in the future.

So, remove this to make the presenter's logic simpler and clearer.

Pick-to: 6.6 6.5
Change-Id: Idffd2cf7e531271e218bb8434a4f3bb287eb4830
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>