Spatial Audio: Fix QAudioListener crash when d->engine = nullptr
Fixes an issue where a QAudioListener with d->engine set to nullptr
could cause the application to crash due to a missing nullptr check.
This situation occurs if one attempts to connect a QAudioListener to
a QAudioEngine which already has a connected QAudioListener.

Fixes: QTBUG-118777
Pick-to: 6.5
Change-Id: I5301dfde264c0fcf8cd6a282915a4819c1f8644a
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit d901e462eb83087ea64610438027652346398f0f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>