Harden color management regression tests
We are seing flaky color management tests on CI with Red Hat Linux. When
inspecting the output images stored during the test run, there are no
significant differences.

This patch changes the logic that stores the actual output images to
store the exact same images as the image comparison algorithm sees. In
addition, a difference image is stored to help investigating flakiness.
The goal is to be able to capture the differences in the stored images
on CI.

This patch also switches to double precision when calculating Y-channel
values in the input images, to make it symmetric with the UV-channels,
and reduces overhead by passing more data by const reference.

In addition, we extend the range of the input pixel values to use the
full 0..255 range in Y, U, and V channels. This makes the test more
challenging when facing non-conformant media. This changes the reference
images.

Change-Id: Icef3d501f6f95379369d2a61117a5b90c2549c33
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 80a9b9d46c46f903e54ae5603ebdd52847398bd9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d0b8636dcae49b2d27f1a808eeb5521c2e659843)