EVRCustomPresenter: fix processing samples by the Scheduler
Fix several issues in how the Scheduler processed the sample
queue:

- Potential data race on accessing to m_scheduledSamples member,
because it was not guarded by the mutex in the `while` condition.

- Potential infinite loop in the scheduler thread caused by
incorrect 0 clock time being returned. This causes the sample
to always be from the future, and it therefore remains queued.
So, the scheduler cannot stop.

Fixes: QTBUG-120266
Pick-to: 6.6 6.5
Change-Id: Icb5a92edf92ab555c9660feeb2a0fce32a037c34
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit cc8492f75c285b58b90c16d8712b19ba181db443)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>