tst_qscreencapturebackend: Change comparison for captured image size
Before this fix, the size of the captured image was compared to the
expected size. The expected size was calculated as the screen size
multiplied by the pixel ratio. This solution works when the pixel ratio
is an integer value, but some Android devices may have different value
(like 2.75). In such case, the expected value may be incorrectly
calculated. It happens because some precision is lost when rounding.

For example on device with 1080x2280 resolution and pixel ratio 2,75
the screen size will be calculated to (392.7272, 829.0909) and rounded
to (393, 829). In such situation, expected size will be wrongly
calculated: 2,75 * QSize(393, 829) = QSize(1080.75, 2279.75).

Because of that it will be better to compare calculated screen size from
captured image rather than calculated expected screen size.

Pick-to: 6.7 6.6 6.5
Fixes: QTBUG-118573
Change-Id: Ie2fc371f2f651ba7754b17cc020e8deedfdf11b3
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>