Fix corner case in 'soft' audio/video sync in the playback engine
There's a corner case when a user changes playbackRate when 'soft' sync
is turned on, it was handled incorrectly, and soft sync data was broken
in result.

The corner case is the main reason for flakiness of
the test case 'playbackRateChanging'.

The fix allows to set stricter checkings in the test cases.

Also, the test case is extended and prettified in the patch.

Pick-to: 6.5
Change-Id: I19e70f3e1bfd66432e8cda8b470c895fde40ad86
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 0cd8de91513125a9bac8f5dd384ddbc972b104a7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>