Android-backend: Set new display on MediaPlayer when Rhi updated
In some cases player can be started before QRhi was set in videoSink.
In such case, surfaceTexture without Rhi will be used for setDisplay.
It may happen that Rhi will be set later, but MediaPlayer will not be
informed about it.

For this case, rhiChanged Signal need to be handled by
qandroidmediaplayer. This allows to update surface when Rhi is set.

This is known issue. Similar problem was already fixed for Camera
Preview in 0389cb306ac17ec9dbbc5ee97ac89659ad4d02b5 commit.

Pick-to: 6.6 6.5
Fixes: QTBUG-118699
Change-Id: I8bded2487a1cad351e4d2a6954a5abe6825d50eb
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>