EVRCustomPresenter: skip samples that came too late
In some cases, the mixer may produce output samples that are already
stale, i.e. their presentation + duration extends beyond the current
clock time. This occurs when the video file's native sample rate
exceeds the actual video frame rendering rate. This is especially
true when HW acceleration is not used, so we spent more time presenting
each video frame.

The current implementation simply schedules _all_ video samples
for presentation, which can lead to problems when the video plays
at a slower speed and reaches the end later than expected.

So, to solve this, just check if the sample time is already passed,
and if yes - discard, i.e. return it to the pool immediately.

Fixes: QTBUG-118587
Pick-to: 6.7 6.6 6.5
Change-Id: I449e4ea7ef7b74a843e12ce7f488bc10c5087f76
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>