Ensure code compatibility 6.5+6.6 with 6.7 due to RotationAngle changes
In the matching CR in the dev branch,
the following changes in API have been introduced:
- Added enum QVideo::RotationAngle
- enum QVideoFrame::RotationAngle is deprecated
- Added overloaded method QVideoFrame::setRotationAngle(QVideo::RotationAngle).

These changes affected many files in different backends.
The idea is to ensure code compatibility of 6.5 and 6.6 with
the dev branch in order to reduce the number of merge conflicts and
save time for backports in the future.

The CR is a cherry-pick of
codereview.qt-project.org/c/qt/qtmultimedia/+/462802
+ the private header video_p.h instead of API changes.
The idea is to make compilable all usages of QVideo::RotationAngle,
including QVideoFrame::setRotationAngle(QVideo::RotationAngle).

Change-Id: Icfbd766e12ee1331de5235a535348b502ec158f0
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 7b28679abe1a33eafe0e091ca19455caaef882ae)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>