Improve the selection of media files in tests
Media selection is used for prechecking existing urls and selecting
a supported one of the given candidates.

Let's check in a separate test case that all used test media files
have been selected, and skip specific tests if the matching video
is not available (as it was before the patch).

Also, the patch improves error dumping of failed media files.

Pick-to: 6.6 6.5
Change-Id: If38c8d9cacf233db75f1469de61dbe320c00f64c
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>