Android: Use the zoom setting for taking photos in FFmpeg
Builder for CaptureRequest was missing setting zoom
([0]SCALER_CROP_REGION) when capturing photos. Because of that, zooming
was ignored for taken pictures.

This commit adds mZoom as QtCamera2 class member and use current value
for taking the photos.

Additionally use mZoom also in start() method. It will be useful for
camera restarts (mZoom will be kept between restarts like mFlashMode).

[0]https://developer.android.com/reference/android/hardware/camera2/CaptureRequest#SCALER_CROP_REGION

Fixes: QTBUG-118309
Pick-to: 6.6 6.5
Change-Id: Ic9c94f4321d1aef1fe6b4a596c22c9ecc2128ad1
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>