QFFmpegMediaPlayer: Fix stop() not resetting position if EndOfMedia
If QFFmpegMediaPlayer::stop() was called after playing a media file all
the way to the end, the playback position of the  would not be reset.
In that scenario, a subsequent call to play() would just reset the
position to 0 instead of starting playback from the beginning as
expected.

Solution:
- Call m_playbackEngine->seek(0) in stop()
- Write new test that covers the above scenario

Fixes: QTBUG-118127
Pick-to: 6.6 6.5
Change-Id: Ibcb313721099c3e53fd6b1b1d52aad712a0c8eb0
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit ad783bb7099fbbc98cf1a27ea8a5910604fc7596)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>