Android: Check if CONTROL_AF_STATE is not null
Fix the crash related with java.lang.NullPointerException.

[0]CaptureResult.get() method can return null if the field is not set.
In documentation of [1]CONTROL_AF_STATE they wrote that "This key is
available on all devices", but it seems that the value can be null
anyway.

This commit validate the afState before using it. Null case is handled
in the same way as in [2]google camera example.

[0]https://developer.android.com/reference/android/hardware/camera2/CaptureResult#get(android.hardware.camera2.CaptureResult.Key%3CT%3E)
[1]https://developer.android.com/reference/android/hardware/camera2/CaptureResult#CONTROL_AF_STATE
[2]https://android.googlesource.com/platform/development/+/40f40d136303c8e399be33d2c55924958268bc03/samples/browseable/Camera2Basic/src/com.example.android.camera2basic/Camera2BasicFragment.java#295

Pick-to: 6.5
Fixes: QTBUG-119089
Fixes: QTBUG-119445
Change-Id: I8b34517ec2b031cdacb4d775af60740016b3e27c
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 2273293ba93d6ec5b8b37f47bc186ed01be2f844)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>