Avoid skipping frames with windows media player
We see problems with jittering/jumpy media playback with the windows
backend. I speculate that the media foundation decoder overwrites
shared textures before or while they are being presented.

This patch attempts to work around this issue by copying the video
texture from the media foundation sample while we still have ownership
to the media sample. This way it does not matter if the display of the
texture happens later, after the media sample has been returned to the
media foundation.

The overhead of creating a backing frame texture and copying the texture
data appears to be low.

Fixes: QTBUG-117099
Pick-to: 6.5
Change-Id: Ib06f0875a19902df2e0c930396529bba4d31a3da
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 78c79152bcfda43187cf77d913d887045e14b6bf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>