Fix PulseAudio sink intermittent hanging and glitching in pull mode
PulseAudio sink in pull mode asks the sound server for the amount of
bytes it wants the client to write on the playback stream. Data was
written from source in chunks of 20 milliseconds, without going over the
requested amount. This however sometimes lead to a state where the
server asks for less bytes than the sink considers as one chunk, which
caused audio hang-ups.

Now the PulseAudio sink will completely write the amount requested by
the sound server. Additionally the audio source is reset when closing
the sink to avoid glitches when playing audio from the same source
later. The timing of chunk writes is adjusted according to the buffer
target length in order to to lower the chance having to write a chunk in
two parts.

Fixes: QTBUG-116519
Pick-to: 6.7 6.6 6.5
Change-Id: I97816c0b7c3c98f3ff6a9d5b1329c616a0f00aa8
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>