Pass HDR info through from QQuickVideoOutput to QSGVideoMaterialRhiShader
We need the maximum luminance during QSGVMRS::updateUniformData so we can
pass it on to QVideoTextureHelper::updateUniformData.

Ideally the logic to compute the max nits based on the QRhiSwapChainHdrInfo
would be shared somewhere, but that's left for a follow up.

We also plumb QRhiSwapChainHdrInfo though directly, but in the future
we might want to join QRhiSwapChainHdrInfo and QRhiSwapChain::Format,
since we only use the latter to check if HDR is enabled when setting
up the shaders.

Done-with: Pavel Dubsky <pavel.dubsky@qt.io>
Change-Id: Id1bcc87a93c9252d90dab8691b939b6b9a30f5bc
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 997364d524a9741e18c97bc8d403726f5f4ec361)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>