Make sure QCamera::errorOccurred is emitted with error in valid range
QCamera's Error enum allows only two values, NoError or CameraError. We
must therefore make sure that QCamera::error() always returns value in
the legal range, and that the QCamera::errorOccurred() signal is always
emitted with a value in the documented range.

On Windows we violated the range for the Error enum because we passed
HRESULT values from the underlying backend through the QCamera error
APIs. This would for example happen if the camera was already busy when
attempting to start the declarative-camera example.

This patch fixes this bug by hardcoding the error code to CameraError,
independently of the internal Windows error code.

Fixes: QTBUG-119236
Pick-to: 6.5 6.6
Change-Id: I9ef4b0a4fd9f32947f8474e41c03821e21d1e627
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>