Fix error property doc for ScreenCapture and WindowCapture
The ScreenCapture::error and WindowCapture::error properties are
enumerations not strings. With this patch, the documentation of the
error properties will be the generic 'enumeration' without specifying
that the type is QScreenCapture::Error or QWindowCapture::Error, but at
least it is not documented as being a string.

Task-number: QTBUG-119236
Pick-to: 6.5
Change-Id: Ic3161f5b7598fad250878d28a219c6b2a3c8a21b
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 8fad1fafc746ff70971413302770d3681dae8f5a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>