Cleanup MFVideoRendererControl in 'windows' backend
In 40793f147be4160fef7ba2d3a781d02de0661a72 we added custom EVR
presenter in order to support HW accelerated video playback (see
EVRCustomPresenterActivate). But initially it was only available
if the surface supported OpenGL texture handles, and only on Windows
Vista or higher. In cases where it was unsupported we fell back to
the old flow (VideoRendererActivate).

Since 2013, when this change was introduced, we have stopped
supporting all Windows versions that do not support EVR, and have improved our custom EVR presenter in order to support rendering on
any RHI backend on Windows.

Thus, the old unoptimized code (VideoRendererActivate, etc.) is
useless now and we can easily remove it to make the rest code more
maintainable.

Pick-to: 6.7 6.6 6.5
Change-Id: I1d1cda13352d8dc70e29f503ee95ebce4979013b
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>