Android-backend: Use the same thread by TextureVideoOutput after reset
7d53ccf2623dfacd4cc8f400729fb78d886924b4 commit fixed the Memory leak
related to not deleted QRhi. QRhiWithThreadGuard was added to keep m_rhi
and m_thread until frame will be released.

Currently there is a problem when frame is kept too long. After
unlocking the screen and going back to app, QRhi is recreated. It is
recreated from the new thread. If the old frame (from previous m_rhi and
previous m_thread) was somehow still kept and it will be released after
that, app will crash with:

    "Cannot make QOpenGLContext current in a different thread"

After mentioned commit, QRhi is recreated from the new thread. The
reason of re-creating AndroidTextureThread was to not handle old pending
frames after reset. As this solution may cause the crash - it needs to
be handled in other way.

This commit stops re-creating texture thread during reset. Instead of
this, there is an additional check if current surfaceTexture is the same
that was used for connection.

Pick-to: 6.7 6.6 6.5
Fixes: QTBUG-118839
Change-Id: I853c3c96b690778435fd592898973f1f478fcea3
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>