qmlvideo: split the project into different qml modules
We want to encourage people to split their qml projects into modules,
which is friendlier for out tooling, and makes better use of out cmake
API.

Since the qmlvideo example was already split into 3 parts
(frequencymonitor, performancemonitor, and the rest of the app)
It made sense to turn those 3 parts into their own qml modules.

This way, we can also make better use new the cmake API, e.g:
qt_add_qml_module(), qt_standard_project_setup(), etc.
In addition to QQuickView::loadFromModule() which was introduced in 6.7.

This patch also makes the example adhere to our guidelines for examples
better, but a follow-up patch is needed in order to make it fully compliant
with our guidelines for examples.

Task-number: QTBUG-113689
Pick-to: 6.7
Change-Id: Icd5cdc56fda8fbcd101e76d6f15ad31609889465
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>