Clean up QPulseAudioSink pull/push logic
PulseAudio sink tick timer and userFeed method have confusing overlap
when it comes to pull mode/push mode. The tick timer is used for
feeding data to the output stream in pull mode, but for setting an
underflow error state in push mode.

This patch renames some variables to disambiguate things, and refactors
the logic in tick timer callback and userFeed method.

Pick-to: 6.7 6.6 6.5
Change-Id: I6c8df50d73e3d0e6f5557f8da141150ad0154ff9
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>