Fully drain PulseAudio stream when stopping the sink
Currently when a PulseAudio stream is closed, pa_stream_drain is called
and the stream is disconnected and deleted, which causes the draining
operation to fail.

This patch moves stream drain to be handled synchronously when calling
stop(), after which close() is called which handles stream disconnect.
The stream is flushed in close() so it can be called for sink
destruction and reset where draining is not wanted.

Task-number: QTBUG-111045
Pick-to: 6.6 6.5
Change-Id: I7a6daf38a410518e98974b807e8f9c8ed9802293
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 9262a8abde7aa2aa5629bc8c382eda2ab91d8e96)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>