Tests: get correct file path for Android in tst_qaudiodecoderbackend
After refactoring done in 8268f5007ed4cdf13ec9dc3250f1acffc56c65ac
commit, path for m_wavFile is not correctly calculated for Android.
QFINDTESTDATA returns this in qrc (":/testdata/test.wav") that is not
the one we are looking for.

Same issue was already some time ago and fixed in
238678e4dd0a45ea13cddbc11dde640a898a1042 commit. For Android we used
temporary file to get correct path.

This patch contains same way of handling the problem for m_wavFile file
in mediafileselector

Pick-to: 6.7 6.6 6.5
Task-number: QTBUG-118572
Change-Id: I05faac8da0a24d77ac026f82f0b4d204a32c8660
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>