Add an opportunity to select screen capture impl by env variable
We need this feature for testing screen capturing on the user's side.
The same we have for window capturing.
In most cases, we will ask users to fall back to 'grabwindow' backend.
Specifically, EglfsScreenCapture implementation may have some
corner cases that we cannot consider now, so this feature
might help us and users.

Pick-to: 6.7 6.6 6.5
Change-Id: I76717a750c54fb931cca15a2bbaec56f9ab9fc3c
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>