Fix bumpy video playback on Windows with FFmpeg using synced textures
We have been facing issues with bumpy video playback when attempting to
keep the decoded textures on the GPU to avoid copying decoded textures
to the CPU. This is seen as stuttering in the video playback, and it
looks like we are switching between new and old frames during playback.

A root cause for this can be that we used texture sharing without
properly synchronizing access to it. This patch fixes this issue by
using synchronized surface sharing as recommended by Microsoft
documentation. This patch also reduces the number of per-frame texture
allocations.

This way we get flicker free rendering during video playback.

Task-number: QTBUG-111815
Task-number: QTBUG-115308
Task-number: QTBUG-111459
Task-number: QTBUG-109213
Pick-to: 6.5 6.6 6.7
Change-Id: I6ed34051f138fc845d60b4208fdbef2ae7d68189
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>