QVideoWindow: Immediately render window on an Expose event
QWindow::requestUpdate() is used to signal that the window
needs to be redrawn (typically because something changed),
but the actual drawing will not happen until the next
event loop iteration. On an Expose event, we need to
immediately draw the window to avoid any delay.
Replace the call to requestUpdate() on an Expose event
with a direct call to render(), ensuring a prompt window
rendering.

Pick-to: 6.6 6.5
Change-Id: Ib74c1ca9706fdc6d92ae8ce659a4fc345edc77d5
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 7318eefeab392d102b9092d649e3f1f3ad12456b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>