Implement audio resampler in the ffmpeg plugin
QFFmpeg::Resampler has been renamed and reworked for inheritance from
QPlatformAudioResampler.
The main purpose of the set of commits is to align the sound played by
QMediaPlayer and QSoundEffect. The classes have different volume
levels and channel mixing factors if we use ffmpeg resampling in one
and doesn't use in another one.

Pick-to: 6.7 6.6 6.5
Task-number: QTBUG-118099
Change-Id: I01cb909de756d5e7cd0f1c40e749f186efd2f649
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>