Fix abrupt termination during static destruction in QtMultimedia
Breaking change: After this change, QtMultimedia requires a
QCoreApplication to access or enumerate media devices (microphone,
speakers, and camera).

During destruction, a Windows application using QtMultimedia could exit
with success return code without having called all static destructors.

The root cause of the issue was that the internal QWindowsMediaDevices
singleton was making outgoing COM-calls during static destruction. This
is not handled well by the COM runtime, and resulted in
NtTerminateProcess being called from within the implementation of
Microsoft IMMDeviceEnumerator::UnregisterEndpointNotificationCallback.
After this point, destructors were never called even if the application
returned 0, indicating success.

This patch fixes this issue by ensuring that the QWindowMediaDevices
instance are deleted when the QCoreApplication is terminated, instead
of waiting until static destruction. If no QCoreApplication exists,
a warning is emitted, and destruction is executed as before, with
possible interruption of static destructors.

Fixes: QTBUG-120198
Pick-to: 6.7 6.6 6.5
Change-Id: I52a2b355052d61333fbff382dc4d4cd059d99d52
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>