Fix regression causing asserts with FFmpeg media player
While fixing bumpy video playback on Windows with FFmpeg, we introduced
two regressions that cause application to assert in debug
(3363496a7f0a0690bc3a2e50172cf6563f198701). These bugs were not
discovered because I was testing with a WARP device, not a proper GPU.

The first issue is that the shared texture between FFmpeg and RHI was
created with 0 mip levels. This causes validation error in the D3D debug
layer, and causes assert when playing the first frame. The fix is to
set 1 as MipLevels.

The second issue is that when opening a second file in media player, we
did not close the shared handle before recreating it. This causes an
assert. The fix is to close the handle first.

Task-number: QTBUG-111815
Task-number: QTBUG-115308
Task-number: QTBUG-111459
Task-number: QTBUG-109213
Pick-to: 6.5 6.6 6.7
Change-Id: Ifd621b174fc68ef50e40cb3ad2d1c4bd410cf15e
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>