Implement audio resampling in QSoundEffect
The resampling in QSoundEffect is needed to align the
audio playback functionality with QMediaPlayer,
otherwise, we get different volume levels when
channels configurations of the source file and the audiodevice
are different.
QMediaPlayer uses ffmpeg resampling, so QSoundEffect is
supposed to use it too.
FFmpeg implementation is to be added in the following commits.

Pick-to: 6.6 6.5
Task-number: QTBUG-118099
Change-Id: I24d0ca394635d53e8fdd4b09f328725dd4f51027
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit f0016c4763361292c3eaf153e184b9b110f44c61)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>