Fix conversion of QImage::Format_RGBA8888_Premultiplied to PixelFormat
QImage::Format_RGBA8888_Premultiplied should be converted to
QVideoFrameFormat::Format_RGBA8888_Premultiplied which is not present
in the list of formats. It should be added in 6.8.

We've got the format on eglfs, when we grab a frame by OpenglCompositor.

The best workaround is conversion to QVideoFrameFormat::Format_RGBX8888.

Pick-to: 6.6 6.5
Change-Id: Ic84d47f62c379e4a476a868b2985b525c0ec7edb
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit bdc9e0a2e30d2a7a7edf0c52630eb8137038cd89)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>