Examples: Fix not visible text in recorder example
The "Capture" and "Settings" texts were not visible on Android devices.
The size of the buttons and default padding caused the text to not fit
inside.

To make the text fit the size of the buttons, clear all paddings.

Pick-to: 6.6 6.5
Fixes: QTBUG-121200
Change-Id: I6c7480237ed4d9ea374929105d019d04b590028a
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit de674e0abfad4df4ea11daccd31a9b77b3010956)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>