Ensure FFmpeg AudioSourceIO COM unitialization happens on worker thread
At construction, QFFmpegAudioInput creates a QFFmpeg::AudioSourceIO
object, and moves it to a worker thread. The AudioSourceIO will then
create a QWindowsResampler which is designed to initialize COM on this
worker thread in its constructor and uninitialize COM in its destructor.
This is needed because the QWindowsResampler uses Windows Media
Foundation.

Although the AudioSourceIO object was moved to a worker thread,
ownership of object lifetime was not transferred. This caused the
QWindowsResampler destructor to uninitialize COM on the main thread
instead of the worker thread.

This patch fixes the issue by moving ownership of object lifetime to the
worker thread.

Amends: 45019c7f9ff9280c5aa2136f4dc60ec8e802d8c3

Pick-to: 6.7 6.6 6.5
Fixes: QTBUG-121495
Change-Id: Ib73518881e2d40d3f999e6ee6ee7565619a48e61
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>