Rename QAudio namespace to QtAudio
With the addition of the QtVideo namespace in Qt 6.7 it becomes
obvious that the existing QAudio namespace is not named according
to the convention for namespaces in Qt.
To fix this, we can rename the QAudio namespace to QtAudio, and
declare the QAudio namespace as an alias for QtAudio. Public APIs
using a QAudio type in their signature need to be exported from a
removed_api.cpp file to maintain binary compatibility.

Two of those are a signal, which leaves one challenge: string-based
connects will have to use the QAudio namespace, in spite of the
documentation, as the signal doesn't exist with a QtAudio parameter.
Add a \note in the signals' documentation.

[ChangeLog][Qt Multimedia][Potentially Source-Incompatible Changes]
The QAudio namespace has been renamed to QtAudio, with QAudio still
being available as an alias. String based connections to
QAudioSink/Source::stateChanged need to continue to use
QAudio::State as the parameter.

Pick-to: 6.7
Change-Id: If49676ee62e048efbed68bb0e4848a5fefaea2f7
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>