Document that Qt Multimedia requires a QCoreApplication
We have previously had examples that demonstrate use of QtMultimedia
APIs without first having created an application object. This is
unfortunate because to fix certain types of bugs, we need to restrict
the lifetime of internal Qt Multimedia state to the lifetime of the Qt
application.

This patch addresses this issue by documenting that a Qt application
object must be created before any of the Qt Multimedia APIs are
accessed, and that if the application object is recreated, multimedia
objects must be recreated as well.

Pick-to: 6.6 6.5
Change-Id: I95645c95ac922c6adb0287d4d02ffb4f385fa60d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
(cherry picked from commit 9cf4494d49170eafb9c179d5304c0a70eaf25b96)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>