Make sure Windows Media Foundation is initialized before using it
With the FFmpeg backend we use Windows Media Foundation to implement the
resampler. The Windows Media Foundation was, however, not initialized
before being used. The consequence of not initializing Windows Media
Foundation first is not completely understood, but we have seen it being
related to performance issues.

This patch initializes the Windows Media Foundation before using its
functionality. In addition, we use the QComHelper class to ensure RAII
initialization of the COM runtime.

Task-number: QTBUG-120026
Pick-to: 6.7 6.6 6.5
Change-Id: I4e1e58229776b42d2b76bfc0828f789e9072fde7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>