Improve performance of retrieving video devices on Windows with FFmpeg
Enumerating USB video devices was slower with the FFmpeg backend on
Windows than with the Windows backend. The reason for this issue was
that the Windows Media Foundation was not initialized before using WMF
APIs to enumerate devices. I suspect that WMF would then have to
initialize and uninitialize WMF for every device, and I observed that
releasing IMFMediaSource instances in the private createCameraDevice
function took over 200 ms.

This patch ensures that Windows Media Foundation is initialized before
accessing media devices. This reduces the per-device enumeration time to
~10 ms on my computer. Additionally, the patch switches to RAII
initialization of the COM runtime, because the COM runtime must be
initialized before accessing Windows Media Foundation APIs.

Fixes: QTBUG-120026
Task-number: QTBUG-120198
Change-Id: I96084689e603bf460708f7e97b2e01e045cb5023
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>