Use regular plugin loader also in multimedia unit tests
This patch switches to regular plugin loading also in unit tests by
introducing a 'mock' media plugin. This plugin is a static plugin, and
does not have to be located in the plugin loader's search path.

This allow us to create the mock plugin by setting an environment
variable instead injecting a mock factory into the
qplatformmediaintegration through private APIs.

The purpose of this change is to simplify how we create the multimedia
backend, and later get better control over its lifetime.

Pick-to: 6.7 6.6 6.5
Change-Id: Ie6cc0ebecc21fb4d81e57cc9f257e4fc82f96e6b
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>