Android-backend: Do not reset surface for camera recording
Currently restarting surface that is used by camera and setting the
new one on MediaRecorder cause preview freeze.

It happens because we destroy AndroidSurfaceTexture that was used for
Camera Preview, but QRhi is not destroyed yet. It will be destroyed
with destruction of last AndroidTextureVideoBuffer (that happen in
feature).
The new surface works (video is correctly recorded), but cannot be
successfully set for Camera Preview (by MediaRecorder.setPreviewDisplay)
that is still using old QRhi.

As it is described in documentation[0], if
setCamera(android.hardware.Camera) is used and the surface has been
already set to the camera, application do not need to call
setPreviewDisplay().

This commit removes code, that was added nine years ago
(43f3d14a99431da5e0ebd29f6c1f27526a6e2d48) as a workaround for some old
devices (with Android 4 that is not longer supported). Currently this
code freezes the preview and seems completely unnecessary.

[0]https://developer.android.com/reference/android/media/MediaRecorder#setPreviewDisplay(android.view.Surface)

Pick-to: 6.6 6.5
Fixes: QTBUG-121714
Change-Id: I2f15a22ed29e3cc7a036d94ff8fc9902687d1bfd
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 4aedcd7c3c1e104216e9d2bbad3cf6a43cda434a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>