SpatialAudio: Resolve some thread safety issues in QAudioEnginePrivate
Previously it was possible for the main thread change which sources of
data were avaialble while the audio thread was generating sounds from
those sources. This changes makes it where sounds and rooms can't be
added or removed while the audio thread is using that state. It also
addes additional checks to make sure we don't end up deferencing
nullptrs.

Change-Id: Iaecade4200da60c074c187c757c1ee56b5a13a10
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>