SpatialAudio: Make sure channel count is always defined
In the case that channelConfig is QAudioFormat::ChannelConfigUnknown, then
we end up not setting any channel count at all and it defaults to 0
which causes issues (no sound). So attempt to set a channelCount from
the devices preferred format instead.

This patch also simplifies the buffer size code below since there is a
function as part of QAudioFormat that accepts a microsecond based
duration the buffer should be.

Change-Id: I035a1655be41feb8580a861a1c3c98ccc5c62fea
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>